pyproject.toml
[project]
name = "support-buddy"
version = "0.1.0"
description = "AI-powered internal support tool for Technical Support Engineers"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40,<1.0",
    "chromadb>=0.4,<1.0",
    "fastapi>=0.100,<1.0",
    "uvicorn>=0.24,<1.0",
    "pydantic>=2.0,<3.0",
    "rich>=13.0,<14.0",
    "typer>=0.9,<1.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
]

[project.scripts]
support-buddy = "src.cli:app"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
strict = true

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"
